+++ GUI-Demo/app/src/main/java/com/mingle/ZiYou/bean/Comment.java
@@ -1,0 +1,7 @@
+package com.mingle.ZiYou.bean;
+
+/**
+ * Created by Kinney on 2016/4/24.
+ */
+public class Comment {
+}
