+++ GUI-Demo/app/src/main/java/com/mingle/ZiYou/dao/CommentDao.java
@@ -1,0 +1,7 @@
+package com.mingle.ZiYou.dao;
+
+/**
+ * Created by Kinney on 2016/4/24.
+ */
+public class CommentDao {
+}
