+++ GUI-Demo/app/src/main/java/com/mingle/ZiYou/dao/PointDao.java
@@ -1,74 +1,26 @@
 package com.mingle.ZiYou.dao;
 
+import com.mingle.ZiYou.bean.Point;
 
-import cn.bmob.v3.BmobObject;
-import cn.bmob.v3.datatype.BmobFile;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
- * Created by jd on 2016/4/14.
+ * Created by Kinney on 2016/4/24.
  */
-public class PointDao extends BmobObject {
-    Integer pid;
-    String pname;
-    Integer sid;
-    BmobFile pmp3cn;
-    BmobFile pmp3en;
-    String plong;
-    String plat;
+public class PointDao {
+    private List<Point> pointList = new ArrayList<Point>();//所有景点列表
 
-    public Integer getPid() {
-        return pid;
+    //通过景区id获得该景区所有景点列表
+    public  List<Point> getPointDaoListBySceneId(int sceneId){
+        return null;
     }
-
-    public void setPid(Integer pid) {
-        this.pid = pid;
+    //通过景点名字查找景点信息
+    public Point getPointByName(String ponitName){
+        return null;
     }
-
-    public String getPname() {
-        return pname;
-    }
-
-    public void setPname(String pname) {
-        this.pname = pname;
-    }
-
-    public Integer getSid() {
-        return sid;
-    }
-
-    public void setSid(Integer sid) {
-        this.sid = sid;
-    }
-
-    public BmobFile getPmp3cn() {
-        return pmp3cn;
-    }
-
-    public void setPmp3cn(BmobFile pmp3cn) {
-        this.pmp3cn = pmp3cn;
-    }
-
-    public BmobFile getPmp3en() {
-        return pmp3en;
-    }
-
-    public void setPmp3en(BmobFile pmp3en) {
-        this.pmp3en = pmp3en;
-    }
-
-    public String getPlong() {
-        return plong;
-    }
-
-    public void setPlong(String plong) {
-        this.plong = plong;
-    }
-
-    public String getPlat() {
-        return plat;
-    }
-
-    public void setPlat(String plat) {
-        this.plat = plat;
+    //通过景点名字获取语音,暂定获取的是url
+    public String getMP3ByPointName(String ponitName){
+        return null;
     }
 }
